Joint work with Marja-Liisa and Francis.
This checkin adds most of the gradation patterns, only appr 
10 patterns have not been dealt with.

The i as consonant following a vowel behaves differently in North and 
Inari Saami, here we do i4 as the consonant underlying the surface i,
and let surface j be underlying j, and surface vowel i be underlying i.
The reason for this is that postvowel j is actually written both i and j
in Inari Saami.

x:y gradation
Here we remove t:đ from x:y, since t:đ participates also in a tk:đh pair,
an alternation type not found for the other x:y pairs. Thus, we added two 
rules, one t:đ gradation and one tk:đh gradation.

The xx:x gradation formerly subsumed  the first part of xx:y, this was now outsourced,
and we were left with pure xx:x.

The xy:wz gradation rule was removed, as it now is the tk:đh rule, and 
interacts with t:đ.

To the yx:zx gradation rule we added a forgotten hc:vc pair.

We then added the following new consonant gradation rules:

"xyy:xz gradation 1"
"xyy:xz gradation 2" 

"xyy:xy gradation 1"
(here, the name should loose it s digit 1, actually, since no part 2 is needed.)

"xˊy:xy gradation"
Here a very long context is invariant, and we delete the letter apostrophe.

"xˊy:xy gradation"
This is the same as the former, but with the i4 consonant in the left context.
TODO: rename.

"nj to nnj lengthening (not strong grade)"
This rule is a special case of lengthening. Phonologically, it is ńń^RC to 0ńń lengthening,
but since ńń is orthographically nnj, we need a convention different from all other
geminates, thus a separate rule.



git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/smn@109789 c7155fb1-f0a7-4240-a2fc-2600b6f42f90