Restored the rule "xˊy:xy gradation" as one rule.
The Cx and Cy sets contain 78 members each (!),
and are thus hard to read. I wanted to make it
easier to read by splitting them in different
rules with constant context, but the result was
a drastic increase in compilation time.

==> thus back to the old version.

What I would like is a rule 
ˊ:0 <=> Cns _ Cns *: %^WG: ;
but in order to get this I must have a separate
WCLEN: mark different from WG, in order to
avoid deleting the ˊ mark in contexts where
the consonant is lengthened. Testing is forthcoming,
this checkin is to restore compilation time.



git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/smn@112865 c7155fb1-f0a7-4240-a2fc-2600b6f42f90