Started using ^SC, stem consonant.
Changed final stem consonant from Cns to Cns: in the rule
 "x:y gradation" 
as it was, it did not work for words with stem consonant deletion
(as e.g. čeve).

Changed the right context of all the Consonant shortening rules.
Fused to   (%^ÁI:|%^ÁE:) , put it before  (%^SFV:) (%^SC:) 
(the order was the opposite).

Also corrected some tests and added explanations.