Content-wise, this checkin is not a change. The rules are the same as for the last check-in. What is changed is that I have ordered the consonant rules into 4 chapters:
* Consonant gradation
* Consonant shortening
* Consonant lengthening
* Other consonant rules

As it was, we had put rules clowe to each other that were or could be
in conflict. We may do so again, but just now I need to see what CG
rules we do and do not have. This check-in thus in order to avoid
conflicts.

Another question is whether we should do it the smj way...