Positive, and not negative context, for the RAMMUD u-stem contract verb: %> [i|i2:i|Cns] Sgm:* %> ; and not %> \[j Vow|o] ;
Also corrected a stupid mistake for the iä:e rule, it was actually two different rules, iä:ee and iä:e, where the former had i:0 and the latter i:e as its first component. Still to be tested.

Vowel copy suffix in imperative (the %^V rule) did not work for u2, i2. Now, with surface :Vx rather than Vx as the vowel to be copied, it should fare better.