This checkin fixes puáhttiđ to poottim.
Changes \Vow: in many rules, \Vow: is not the same as 
[Cns|.#.|#], which we should use as symbol for onset instead.
Tne \Vow: did not accept uáđđiđ, as beginning of string 
(.#.) in .#.uáđđ is not covered by \Vow:, which means
"a segment that is not Vow. And .#. is not a segment.

Added trigger VBACK for uá to oo, and added it to rules
"Monophthongisation rule uá:o0 part 2: Vowel shortening for uá"
and
"Monophthongisation rule u:o and root vowel change u:o, special rule for olmooš and contract verbs"



git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/smn@120685 c7155fb1-f0a7-4240-a2fc-2600b6f42f90