Added u2:u to Vx Vy def lists of vowel shortening rule .
Changed stem vowel from Vow to Vow: for eRV to iä rule, the stem vowel was in this case u2:u.

Added line
        [Cns|.#.]   _  %^RV:  Cns:+ u2:u %^CLEN: %^VBACK: %^RVSH: %>  ; !RULE RV ! 
to rule "Root vowel change a:o before suffix i and for contract verbs".
The form was anneeđ to onnu-, and this set of triggers was missing.

Also Added ^RV to right context of diphtongisations e to iä, to make it work also for short vowels.



git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/smn@132057 c7155fb1-f0a7-4240-a2fc-2600b6f42f90