This is a debugging checkin, the file is not neccessarily in synch with lexc.
The goal of the cleanup is to get the order we have specified for ^EA and ^CSH 
(the order mentioned). This has often been the case, but not always, and the 
missing inconcistency has given rise to problems.

Now, all rules are rewritten with the order mentioned, as are lexc. The problem
here is that the file does not compile, with the error message 

error for object 0x100800e00: incorrect checksum for freed object - object was probably modified after being freed


git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/smn@137147 c7155fb1-f0a7-4240-a2fc-2600b6f42f90