Fixed the following conflict (thanks to Lene for pointing it out):
 "Stem vowel â to a in Pl Nom -h"
*** Warning: Unresolved <= conflict with respect to 'â:a' vs. 'â:0'
   between "Stem vowel â to a in Pl Nom -h"
       and "Word final vowel deletion"

**** Conflicting pair(s) of contexts: 
 :Cns (^RC:) _ Triggers:* > h [l | .#. | >] ;
  _ ^FVD: ;

The problem was the right context of Word final vowel deletion,
its sole content was the trigger ^FVD, which of course was
a member of the set Triggers.

Now, both rules are changed. At the moment, ^FVD is found only
in front of .#., should this situation change, the rule should
be changed as well.