Several changes
Added set ADV-PO for ambiguous words +Adv +Po. This rather than referring to Po in the rule, since the word may be wrongly disambiguated as Adv due to the missing Gen we want to correct.

Translated ILL-V set.

Added (A Gen) to PRE-NP-HEAD set, since smn differs from sme in not only having A Attr but also A Gen.

3 new rule tags:
&msyn-predagr-pl3-sg3 ;
&msyn-po-placc-plgen ;
&msyn-obj-acc-ill ;
... and rules to identify them.
Especially the last ones should be checked for false alarms.