The experimental &easteregg checkin
This was done:
in root.lex:
- I added the trigger word nuvviDgrammarchecker to the Root lexicon in root.lexc

in grammarchecker.cg3:
- I added a tag &easteregg initially in the file
- I added a rule that flagged the word nuvviDgrammarchecker with the tag @easteregg and then removed both, and replaced them with today's date (TODO: make this aware of time)

In errors.source.xml:
I added a message for te &easteregg.

Now, if the smn.fst would be able to return the compilation date when prompted for it, we would have a working easteregg. The unacceptable quickfix would be to update 19.1.23 each time we uoload the file, clearly absurd.

I take it the nuvviDspeller mechanism from the speller may be carried over somehow.