Intermediate python script for release grammarchecker
This file makes grammarchecker-release.grc from grammarchecker.grc. This presupposes that rules *not* scheduled for release are marked with "ADD:x" whereas releasing a rule means removing the x (default is thus that the rule is not released).

TODO further down: Integrate this in the makefile setup so that a file grammarchecker-release.cg3 is built automatically. In order to make this genral, the sme files should be renamed as follows:
- grammachecker.cg3 > grammarchecker-archive.cg3
- grammarchecker-release.cg3 > grammarchecker.cg3
- ... and the make setup will then treat grammarchecker-release.cg3 as a derived file (perhaps even as grammarchecker-release.bin) for the purpose of the scripts in the modes catalogue.