Several changes, mainy restricting things
Preparing for release. Now removed all x marks (ADD:xmsyn -> ADD:msyn), the old marking made the grammarchecker into an internference one , with only 22 rules. Now we have INDCLUDE RULE as default, and exclude rules as we see them misbehave.

Now   UNITQUANTOR-ADV and MASSQUANTOR-ADV are only maŋgâ, no matter in what form, no need to be in plural, the word may be wrong in the sentence.

DANGEROUSNOUN split into V and Adv lookalikes.

Change tag &msyn-qucompl-sg-pl to &msyn-quant-gen-nom  but then reintroduce the original tag &msyn-qucompl-sg-pl. The old rule targeted quantifier complement (many THINGS) and get the new qucompl tag, the new rule targets the quantifier (MANY things) and get the (old) quant tag.

For all the obj-acc-ill  rules, we ensure the triggering illative verb does not have another ill argument.

For the msyn-nm-acc-par rule, we exclude ordinals as triggers, ensuring chapter numbers of the type "5. Conclusion" do not trigger number phrase alarms.

Redid the quadv-sg-pl and qucompl-sg-pl rules as explained for the tags above.

rewritten the real-love-lope rule, it rested on a mistunderstanding, now restricted to e-sentences where it could be normal case had it not been for the "permission" "ten" homonymy.