Several changes, mostly vowel rules
"Root vowel o:u for contract verbs and Root vowel change o:u before suffix i"

This rule (o:u)
[a|o|i|u] (%^RV:) Cns:+ _ (%^SV:)    %^EA:     ([%^WG:|%^CSH:]) (%^CLEN:) ([%^RLEN:|%^RVSH:]) (%^SLEN:) [%>|.#.] ;

used to be in a seperate rule but was rused with the main o:u rule and adjusted to one rule.

WG, CSH and CSH were split several places.

"ää to áá and á0 in Illative and Sg3"
# was added as left context (allowing compounds)

CSH CLEN was hanged from either-or to possibly either, neither or both (here and in several places, this always holds).

"ye to uá, part 2"
fused with stem vowel e... to
"e to á in ye to uá, part 2 and Stem vowel e to á for other root vowel"

New sentence fused:
[o|i|u] (%^RV:) Cns:+ _ (%^SV:)    %^EA: ([%^CSH:|%^CLEN:|%^WG:]) (%^CLEN:) ([%^RLEN:|%^RVSH:]) (%^SLEN:) [%>|.#.] ;

The point is to treat all instances of e:á in the same rule.

"Stem vowel e to á for other root vowel"
This rule was deleted (fused with previous)

Cosmetic change several places:
([%^CSH:|%^WG:]) to ([%^WG:|%^CSH:])  just to have same order everywhere

"Stem â deleting"
This rule is cometing with
"Word final vowel deletion"

sinde the latter deletes not only final but also stem vowels. The former refers to ii suffix strings, the latter to the FVD trigger. Thus, some dim lexica use FVD, others do not. Unifying would be a good idea, but the double system works (sigh).

Also added tests.

"Suffix i5:0 and i5ä:e"
revised to
"Suffix i5:0 for loanwords"

(the i5ä to e fixed elsewhere). Changing i5 is somewhat awkward, since the x5 notation was for invariant x, we would thus like to revise this.